Flip the inheritance chain between annotation and attribute drivers for the extensions
