+++ src/SoftDeleteable/Mapping/Driver/Annotation.php
@@ -9,61 +9,15 @@
 
 namespace Gedmo\SoftDeleteable\Mapping\Driver;
 
-use Gedmo\Exception\InvalidMappingException;
-use Gedmo\Mapping\Annotation\SoftDeleteable;
-use Gedmo\Mapping\Driver\AbstractAnnotationDriver;
-use Gedmo\SoftDeleteable\Mapping\Validator;
+use Gedmo\Mapping\Driver\AnnotationDriverInterface;
 
 /**
- * This is an annotation mapping driver for SoftDeleteable
- * behavioral extension. Used for extraction of extended
- * metadata from Annotations specifically for SoftDeleteable
- * extension.
- *
- * @author Gustavo Falco <comfortablynumb84@gmail.com>
- * @author Gediminas Morkevicius <gediminas.morkevicius@gmail.com>
+ * Mapping driver for the soft-deletable extension which reads extended metadata from annotations on a soft-deletable class.
  *
  * @deprecated since gedmo/doctrine-extensions 3.16, will be removed in version 4.0.
  *
  * @internal
  */
-class Annotation extends AbstractAnnotationDriver
+class Annotation extends Attribute implements AnnotationDriverInterface
 {
-    /**
-     * Annotation to define that this object is loggable
-     */
-    public const SOFT_DELETEABLE = SoftDeleteable::class;
-
-    public function readExtendedMetadata($meta, array &$config)
-    {
-        $class = $this->getMetaReflectionClass($meta);
-        // class annotations
-        if (null !== $class && $annot = $this->reader->getClassAnnotation($class, self::SOFT_DELETEABLE)) {
-            $config['softDeleteable'] = true;
-
-            Validator::validateField($meta, $annot->fieldName);
-
-            $config['fieldName'] = $annot->fieldName;
-
-            $config['timeAware'] = false;
-            if (isset($annot->timeAware)) {
-                if (!is_bool($annot->timeAware)) {
-                    throw new InvalidMappingException('timeAware must be boolean. '.gettype($annot->timeAware).' provided.');
-                }
-                $config['timeAware'] = $annot->timeAware;
-            }
-
-            $config['hardDelete'] = true;
-            if (isset($annot->hardDelete)) {
-                if (!is_bool($annot->hardDelete)) {
-                    throw new InvalidMappingException('hardDelete must be boolean. '.gettype($annot->hardDelete).' provided.');
-                }
-                $config['hardDelete'] = $annot->hardDelete;
-            }
-        }
-
-        $this->validateFullMetadata($meta, $config);
-
-        return $config;
-    }
 }
