+++ src/SoftDeleteable/Mapping/Driver/Attribute.php
@@ -9,17 +9,63 @@
 
 namespace Gedmo\SoftDeleteable\Mapping\Driver;
 
+use Gedmo\Exception\InvalidMappingException;
 use Gedmo\Mapping\Annotation\SoftDeleteable;
-use Gedmo\Mapping\Driver\AttributeDriverInterface;
+use Gedmo\Mapping\Driver\AbstractAnnotationDriver;
+use Gedmo\SoftDeleteable\Mapping\Validator;
 
 /**
- * This is an attribute mapping driver for SoftDeleteable
- * behavioral extension. Used for extraction of extended
- * metadata from attributes specifically for SoftDeleteable
- * extension.
+ * Mapping driver for the soft-deletable extension which reads extended metadata from attributes on a soft-deletable class.
+ *
+ * @author Gustavo Falco <comfortablynumb84@gmail.com>
+ * @author Gediminas Morkevicius <gediminas.morkevicius@gmail.com>
  *
  * @internal
  */
-final class Attribute extends Annotation implements AttributeDriverInterface
+class Attribute extends AbstractAnnotationDriver
 {
+    /**
+     * Mapping object for the soft-deletable extension.
+     */
+    public const SOFT_DELETEABLE = SoftDeleteable::class;
+
+    public function readExtendedMetadata($meta, array &$config)
+    {
+        $class = $this->getMetaReflectionClass($meta);
+
+        // class annotations
+        if (null !== $class && $annot = $this->reader->getClassAnnotation($class, self::SOFT_DELETEABLE)) {
+            \assert($annot instanceof SoftDeleteable);
+
+            $config['softDeleteable'] = true;
+
+            Validator::validateField($meta, $annot->fieldName);
+
+            $config['fieldName'] = $annot->fieldName;
+
+            $config['timeAware'] = false;
+
+            if (isset($annot->timeAware)) {
+                if (!is_bool($annot->timeAware)) {
+                    throw new InvalidMappingException('timeAware must be boolean. '.gettype($annot->timeAware).' provided.');
+                }
+
+                $config['timeAware'] = $annot->timeAware;
+            }
+
+            $config['hardDelete'] = true;
+
+            if (isset($annot->hardDelete)) {
+                if (!is_bool($annot->hardDelete)) {
+                    throw new InvalidMappingException('hardDelete must be boolean. '.gettype($annot->hardDelete).' provided.');
+                }
+
+                $config['hardDelete'] = $annot->hardDelete;
+            }
+        }
+
+        $this->validateFullMetadata($meta, $config);
+
+        return $config;
+    }
 }
