+++ src/Timestampable/Mapping/Driver/Annotation.php
@@ -9,88 +9,15 @@
 
 namespace Gedmo\Timestampable\Mapping\Driver;
 
-use Gedmo\Exception\InvalidMappingException;
-use Gedmo\Mapping\Annotation\Timestampable;
-use Gedmo\Mapping\Driver\AbstractAnnotationDriver;
+use Gedmo\Mapping\Driver\AnnotationDriverInterface;
 
 /**
- * This is an annotation mapping driver for Timestampable
- * behavioral extension. Used for extraction of extended
- * metadata from Annotations specifically for Timestampable
- * extension.
- *
- * @author Gediminas Morkevicius <gediminas.morkevicius@gmail.com>
+ * Mapping driver for the timestampable extension which reads extended metadata from annotations on a timestampable class.
  *
  * @deprecated since gedmo/doctrine-extensions 3.16, will be removed in version 4.0.
  *
  * @internal
  */
-class Annotation extends AbstractAnnotationDriver
+class Annotation extends Attribute implements AnnotationDriverInterface
 {
-    /**
-     * Annotation field is timestampable
-     */
-    public const TIMESTAMPABLE = Timestampable::class;
-
-    /**
-     * List of types which are valid for timestamp
-     *
-     * @var string[]
-     */
-    protected $validTypes = [
-        'date',
-        'date_immutable',
-        'time',
-        'time_immutable',
-        'datetime',
-        'datetime_immutable',
-        'datetimetz',
-        'datetimetz_immutable',
-        'timestamp',
-        'vardatetime',
-        'integer',
-    ];
-
-    public function readExtendedMetadata($meta, array &$config)
-    {
-        $class = $this->getMetaReflectionClass($meta);
-        // property annotations
-        foreach ($class->getProperties() as $property) {
-            if ($meta->isMappedSuperclass && !$property->isPrivate()
-                || $meta->isInheritedField($property->name)
-                || isset($meta->associationMappings[$property->name]['inherited'])
-            ) {
-                continue;
-            }
-            if ($timestampable = $this->reader->getPropertyAnnotation($property, self::TIMESTAMPABLE)) {
-                $field = $property->getName();
-                if (!$meta->hasField($field)) {
-                    throw new InvalidMappingException("Unable to find timestampable [{$field}] as mapped property in entity - {$meta->getName()}");
-                }
-                if (!$this->isValidField($meta, $field)) {
-                    throw new InvalidMappingException("Field - [{$field}] type is not valid and must be 'date', 'datetime' or 'time' in class - {$meta->getName()}");
-                }
-                if (!in_array($timestampable->on, ['update', 'create', 'change'], true)) {
-                    throw new InvalidMappingException("Field - [{$field}] trigger 'on' is not one of [update, create, change] in class - {$meta->getName()}");
-                }
-                if ('change' === $timestampable->on) {
-                    if (!isset($timestampable->field)) {
-                        throw new InvalidMappingException("Missing parameters on property - {$field}, field must be set on [change] trigger in class - {$meta->getName()}");
-                    }
-                    if (is_array($timestampable->field) && isset($timestampable->value)) {
-                        throw new InvalidMappingException('Timestampable extension does not support multiple value changeset detection yet.');
-                    }
-                    $field = [
-                        'field' => $field,
-                        'trackedField' => $timestampable->field,
-                        'value' => $timestampable->value,
-                    ];
-                }
-                // properties are unique and mapper checks that, no risk here
-                $config[$timestampable->on][] = $field;
-            }
-        }
-
-        return $config;
-    }
 }
